10.6 Serialization Groups
Squash the exception in our query: is_published cannot be null. Oh, that makes sense: the user isn't sending is_published... so nobody is setting it. And it's set to not null in the database. No worries: default the property to false.

By the way, if you're using Symfony 4.3, instead of a Doctrine error, you may have gotten a validation error in previous commit. That's due to a new feature where Doctrine database constraints can automatically be used to add validation. So, if you see a validation error, awesome!

Set is_published property to false instead of null set by default by Doctrine

Try to execute POST collection again. It works! We have exactly the input fields and output fields that we want. The isPublished field isn't exposed at all in our API, but it is being set in the background.

Next, let's learn a few more serialization tricks - like how to control the field name and how to handle constructor arguments.