11.4 @SerializedName & Constructor Args
Constructor Argument can Change Validation Errors

But there is one edge case. Pretend that we're creating a new CheeseListing and we forget to send the title field entirely - like, we have a bug in our JavaScript code. Hit Execute.

We do get back a 400 error... which is perfect: it means that the person making the request has something wrong with their request. But, the hydra:title isn't very clear:

  An error occurred

Fascinating! The hydra:description is way more descriptive... actually a bit too descriptive - it shows off some internal things about our API... that I maybe don't want to make public. At least the trace won't show up on production. Showing these details inside hydra:description might be ok with you... But if you want to avoid this, you ned to rely on validation, which is a topic that we'll talk about in a few minutes.